Moved Content-Item classes to Lti\v1
