+++ LtiLibrary.NetCore.Tests/LtiRequestShould.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Net.Http;
 using LtiLibrary.NetCore.Common;
-using LtiLibrary.NetCore.ContentItems;
 using LtiLibrary.NetCore.Extensions;
 using LtiLibrary.NetCore.Lis.v1;
 using LtiLibrary.NetCore.Lti.v1;
