+++ LtiLibrary.NetCore/Lti/v1/LtiLink.cs
@@ -1,6 +1,6 @@
 ﻿using LtiLibrary.NetCore.Common;
 
-namespace LtiLibrary.NetCore.ContentItems
+namespace LtiLibrary.NetCore.Lti.v1
 {
     /// <summary>
     /// Represents an IMS LtiLink ContentItem.
